requirements.txt
google-generativeai
streamlit
streamlit-chat
python-dotenv
langchain